+++ src/software/ai/hl/stp/tactic/assigned_skill/BUILD
@@ -9,3 +9,12 @@ cc_library(
         "//software/ai/hl/stp/tactic",
     ],
 )
+
+cc_library(
+    name = "named_skill_tactics",
+    hdrs = ["named_skill_tactics.h"],
+    deps = [
+        ":assigned_skill_tactic",
+        "//software/ai/hl/stp/skill/pivot_kick:pivot_kick_skill",
+    ]
+)
